# Use esbuild for tests and build