--- a/jest.config.js
+++ b/jest.config.js
@@ -1,12 +1,5 @@
 module.exports = {
-  "roots": [
-    "<rootDir>/src"
-  ],
-  "testMatch": [
-    "**/__tests__/**/*.+(ts|tsx|js)",
-    "**/?(*.)+(spec|test).+(ts|tsx|js)"
-  ],
-  "transform": {
-    "^.+\\.(ts|tsx)$": "ts-jest"
+  transform: {
+    '^.+\\.tsx?$': 'esbuild-jest',
   },
 }