--- a/package.json
+++ b/package.json
@@ -13,22 +13,19 @@
   "license": "MIT",
   "devDependencies": {
     "@cloudflare/workers-types": "^2.2.2",
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-node-resolve": "^11.1.0",
-    "@rollup/plugin-typescript": "^8.2.1",
     "@types/is-odd": "^3.0.0",
-    "@types/jest": "^26.0.23",
+    "@types/jest": "^26.0.24",
+    "esbuild": "^0.13.4",
+    "esbuild-jest": "^0.5.0",
     "isomorphic-fetch": "^3.0.0",
     "jest": "^26.6.3",
     "prettier": "^1.19.1",
-    "rollup": "^2.36.1",
-    "rollup-plugin-copy": "^3.3.0",
+    "rollup": "^2.58.0",
+    "rollup-plugin-esbuild": "^4.5.0",
     "rollup-plugin-terser": "^7.0.2",
-    "ts-jest": "^26.5.6",
-    "tslib": "^2.2.0",
-    "typescript": "^4.2.4"
+    "typescript": "^4.4.3"
   },
   "dependencies": {
-    "is-even": "^1.0.0"
+    "is-odd": "^3.0.1"
   }
 }