--- a/rollup.config.js
+++ b/rollup.config.js
@@ -1,10 +1,5 @@
+import esbuild from 'rollup-plugin-esbuild'
 import { terser } from 'rollup-plugin-terser'
-// plugin-node-resolve and plugin-commonjs are required for a rollup bundled project
-// to resolve dependencies from node_modules. See the documentation for these plugins
-// for more details.
-import { nodeResolve } from '@rollup/plugin-node-resolve'
-import commonjs from '@rollup/plugin-commonjs'
-import typescript from '@rollup/plugin-typescript'
 
 export default {
   input: 'src/index.ts',
@@ -14,5 +9,11 @@ export default {
     file: 'dist/index.mjs',
     sourcemap: true,
   },
-  plugins: [typescript(), commonjs(), nodeResolve({ browser: true }), terser()],
+  plugins: [
+    esbuild({
+      // required for resolving modules
+      experimentalBundling: true,
+    }),
+    terser(),
+  ],
 }