Just a proposal, no worries if you don't think it's suitable for this project :)

Esbuild runs faster, simplifies the configuration a bit and wouldn't compile the project until (#8) was fixed. The node resolve plugin didn't seem to detect this issue.

### npm run test

before
`npm run test  3.26s user 0.68s system 125% cpu 3.140 total`

after (using esbuild)
`npm run test  2.20s user 0.67s system 107% cpu 2.669 total`

### npm run build

before
`npm run build  3.21s user 0.56s system 173% cpu 2.174 total`

after
`npm run build  1.48s user 0.34s system 213% cpu 0.855 total`