@johnrees 2021-10-11
perhaps experimental isn't the way to go yet. I added an issue to the rollup plugin that I found with injecting env here https://github.com/egoist/rollup-plugin-esbuild/issues/314